+++ src/Response/ListCompanyTeams.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Easir\SDK\Response;
+
+use Easir\SDK\Response;
+
+/**
+ * Class ListCompanyTeams
+ * @package Easir\SDK\Response
+ */
+class ListCompanyTeams extends Response
+{
+    protected $collections = ['data' => 'team'];
+
+    public $data, $pagination;
+}